backend_django/requirements.txt
# TrakNor CMMS Backend - Django 5
Django==5.0.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
djangorestframework-camel-case==1.4.2
psycopg[binary]==3.1.17
django-environ==0.11.2
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==23.5
Pillow==10.2.0
whitenoise==6.6.0
gunicorn==21.2.0

# Dev/Test dependencies
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
black==23.12.1
ruff==0.1.11
django-debug-toolbar==4.2.0